refactor(ui): extract command palette commands to src/ui/commands.ts
- buildCommandPaletteCommands(ctx) returns CommandItem[]
- Removes ~200 lines from AppInner.tsx
- Clean dependency injection via CommandBuilderContext
